Prevent admin from deleting supplier when there are products under that supplier

+++ ShoppingCartFrontEnd/src/main/java/com/niit/controller/SupplierController.java
@@ -15,8 +15,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.niit.shoppingcart.dao.ProductDAO;
 import com.niit.shoppingcart.dao.SupplierDAO;
 import com.niit.shoppingcart.domain.Category;
+import com.niit.shoppingcart.domain.Product;
 import com.niit.shoppingcart.domain.Supplier;
 
 @Controller
@@ -24,13 +26,14 @@ public class SupplierController {
 
 	private static Logger log = LoggerFactory.getLogger(SupplierController.class);
 
-	
 	@Autowired
 	HttpSession session;
 	@Autowired
 	SupplierDAO supplierDAO;
 	@Autowired
 	Supplier supplier;
+	@Autowired ProductDAO productDAO;
+	@Autowired Product product;
 
 	// crud supplier
 	@RequestMapping("/manage_supplier_add")
@@ -41,16 +44,16 @@ public class SupplierController {
 		supplier.setId(id);
 		supplier.setName(name);
 		supplier.setDescription(description);
-		
+
 		mv.addObject("isAdminClickedSuppliers", "true");
 		mv.addObject("isAdmin", "true");
 
 		if (supplierDAO.getSupplierById(id) != null) {
-			mv.addObject("sMessage", "Supplier already exists with id : "+id);
+			mv.addObject("sMessage", "Supplier already exists with id : " + id);
 			return mv;
 		} else {
 			supplierDAO.save(supplier);
-			mv.addObject("sMessage", "Supplier creation success with id : "+id);
+			mv.addObject("sMessage", "Supplier creation success with id : " + id);
 
 		}
 		// Before calling save method, check whether supplier_id already exists
@@ -74,11 +77,19 @@ public class SupplierController {
 	// attach data to url we use @PathVariable
 	@RequestMapping("/manage-supplier-delete/{id}")
 	public ModelAndView deleteSupplier(@PathVariable("id") String id) {
-		
+
 		log.debug("Starting of delete Supplier");
 		log.info("You are about to delete a supplier with id : " + id);
-		
+
 		ModelAndView mv = new ModelAndView("redirect:/manageSuppliers");
+
+		// Check whether products are there for this category or not
+		int noOfProducts = productDAO.getAllProductsBySupplierId(id).size();
+		if (noOfProducts != 0) {
+			log.debug("Few products are there by this seller, you cannot delete!");
+			session.setAttribute("supplierMessage", "There are " + noOfProducts + " products under this " + id + " seller, you cannot delete!");
+			return mv;
+		}
 		if (supplierDAO.delete(id) == true) {
 			mv.addObject("message", "Successfullly deleted");
 		} else {
@@ -88,25 +99,25 @@ public class SupplierController {
 
 		return mv;
 	}
-	
+
 	@RequestMapping("/manage-supplier-edit/{id}")
 	public ModelAndView editSupplier(@PathVariable("id") String id) {
 		log.debug("Starting of editSupplier");
 		log.info("You are about to edit a supplier with id : " + id);
-		
+
 		supplier = supplierDAO.getSupplierById(id);
-		
-		//Selected category details we have to store in another instance
-		//i.e., ModelAndView instance
+
+		// Selected category details we have to store in another instance
+		// i.e., ModelAndView instance
 		ModelAndView mv = new ModelAndView("redirect:/manageSuppliers");
 		mv.addObject("selectedSupplier", supplier);
 		session.setAttribute("selectedSupplier", supplier);
-		
+
 		log.debug("Ending of editSupplier");
 
 		return mv;
 	}
-	
+
 	@PostMapping("/manage_supplier_update")
 	public ModelAndView updateSupplier(@RequestParam("cId") String id, @RequestParam("cName") String name,
 			@RequestParam("cDescription") String description) {
@@ -128,7 +139,6 @@ public class SupplierController {
 			mv.addObject("cMessage", "Supplier updated success with id : " + id);
 
 		}
-
 
 		// Before calling save method, check whether supplier_id already exists
 		// in db
